added markdown-it-configurator made basic- and full-markdown-renderer smaller
