+++ src/components/editor/app-bar/help-button/cheatsheet.tsx
@@ -1,9 +1,9 @@
-import MarkdownIt from 'markdown-it'
 import markdownItContainer from 'markdown-it-container'
-import React, { useCallback } from 'react'
+import React from 'react'
 import { Table } from 'react-bootstrap'
 import { Trans, useTranslation } from 'react-i18next'
 import { BasicMarkdownRenderer } from '../../../markdown-renderer/basic-markdown-renderer'
+import { basicMarkdownRender } from '../../../markdown-renderer/markdown-it-configurator'
 import { createRenderContainer, validAlertLevels } from '../../../markdown-renderer/markdown-it-plugins/alert-container'
 import { HighlightedCode } from '../../../markdown-renderer/replace-components/highlighted-fence/highlighted-code/highlighted-code'
 import './cheatsheet.scss'
@@ -31,11 +31,10 @@ export const Cheatsheet: React.FC = () => {
     `:::info\n${t('editor.help.cheatsheet.exampleAlert')}\n:::`
   ]
 
-  const markdownItPlugins = useCallback((md: MarkdownIt) => {
-    validAlertLevels.forEach(level => {
-      md.use(markdownItContainer, level, { render: createRenderContainer(level) })
-    })
-  }, [])
+  const md = basicMarkdownRender()
+  validAlertLevels.forEach(level => {
+    md.use(markdownItContainer, level, { render: createRenderContainer(level) })
+  })
 
   return (
     <Table className="table-condensed table-cheatsheet">
@@ -53,7 +52,7 @@ export const Cheatsheet: React.FC = () => {
                 <BasicMarkdownRenderer
                   content={code}
                   wide={false}
-                  onConfigureMarkdownIt={markdownItPlugins}
+                  markdownIt={md}
                 />
               </td>
               <td className={'markdown-body'}>
