+++ src/components/markdown-renderer/full-markdown-renderer.tsx
@@ -1,41 +1,13 @@
-import markdownItTaskLists from '@hedgedoc/markdown-it-task-lists'
-import yaml from 'js-yaml'
-import MarkdownIt from 'markdown-it'
-import anchor from 'markdown-it-anchor'
-import markdownItContainer from 'markdown-it-container'
-import frontmatter from 'markdown-it-front-matter'
-import mathJax from 'markdown-it-mathjax'
-import plantuml from 'markdown-it-plantuml'
-import markdownItRegex from 'markdown-it-regex'
-import toc from 'markdown-it-toc-done-right'
 import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react'
 import { Alert } from 'react-bootstrap'
 import { Trans } from 'react-i18next'
-import { useSelector } from 'react-redux'
 import { TocAst } from '../../external-types/markdown-it-toc-done-right/interface'
-import { ApplicationState } from '../../redux'
 import { InternalLink } from '../common/links/internal-link'
 import { ShowIf } from '../common/show-if/show-if'
-import { slugify } from '../editor/table-of-contents/table-of-contents'
 import { RawYAMLMetadata, YAMLMetaData } from '../editor/yaml-metadata/yaml-metadata'
 import { BasicMarkdownRenderer } from './basic-markdown-renderer'
-import { createRenderContainer, validAlertLevels } from './markdown-it-plugins/alert-container'
-import { highlightedCode } from './markdown-it-plugins/highlighted-code'
-import { LineMarkers, lineNumberMarker } from './markdown-it-plugins/line-number-marker'
-import { plantumlError } from './markdown-it-plugins/plantuml-error'
-import { replaceAsciinemaLink } from './regex-plugins/replace-asciinema-link'
-import { replaceGistLink } from './regex-plugins/replace-gist-link'
-import { replaceLegacyGistShortCode } from './regex-plugins/replace-legacy-gist-short-code'
-import { replaceLegacySlideshareShortCode } from './regex-plugins/replace-legacy-slideshare-short-code'
-import { replaceLegacySpeakerdeckShortCode } from './regex-plugins/replace-legacy-speakerdeck-short-code'
-import { replaceLegacyVimeoShortCode } from './regex-plugins/replace-legacy-vimeo-short-code'
-import { replaceLegacyYoutubeShortCode } from './regex-plugins/replace-legacy-youtube-short-code'
-import { replacePdfShortCode } from './regex-plugins/replace-pdf-short-code'
-import { replaceQuoteExtraAuthor } from './regex-plugins/replace-quote-extra-author'
-import { replaceQuoteExtraColor } from './regex-plugins/replace-quote-extra-color'
-import { replaceQuoteExtraTime } from './regex-plugins/replace-quote-extra-time'
-import { replaceVimeoLink } from './regex-plugins/replace-vimeo-link'
-import { replaceYouTubeLink } from './regex-plugins/replace-youtube-link'
+import { fullMarkdownRender } from './markdown-it-configurator'
+import { LineMarkers } from './markdown-it-plugins/line-number-marker'
 import { AbcReplacer } from './replace-components/abc/abc-replacer'
 import { AsciinemaReplacer } from './replace-components/asciinema/asciinema-replacer'
 import { CsvReplacer } from './replace-components/csv/csv-replacer'
@@ -106,8 +78,6 @@ export const FullMarkdownRenderer: React.FC<FullMarkdownRendererProps & Addition
 
   const [yamlError, setYamlError] = useState(false)
 
-  const plantumlServer = useSelector((state: ApplicationState) => state.config.plantumlServer)
-
   const rawMetaRef = useRef<RawYAMLMetadata>()
   const firstHeadingRef = useRef<string>()
   const documentElement = useRef<HTMLDivElement>(null)
@@ -118,7 +88,7 @@ export const FullMarkdownRenderer: React.FC<FullMarkdownRendererProps & Addition
   const tocAst = useRef<TocAst>()
   usePostTocAstOnChange(tocAst, onTocChange)
 
-  const extractInnerText = useCallback((node: ChildNode) => {
+  const extractInnerText = useCallback((node: ChildNode): string => {
     let innerText = ''
     if (node.childNodes && node.childNodes.length > 0) {
       node.childNodes.forEach((child) => { innerText += extractInnerText(child) })
@@ -139,90 +109,15 @@ export const FullMarkdownRenderer: React.FC<FullMarkdownRendererProps & Addition
     }
   }, [content, extractInnerText, onFirstHeadingChange])
 
-  const configureMarkdownIt = useCallback((md: MarkdownIt): void => {
-    if (onMetaDataChange) {
-      md.use(frontmatter, (rawMeta: string) => {
-        try {
-          const meta: RawYAMLMetadata = yaml.safeLoad(rawMeta) as RawYAMLMetadata
-          setYamlError(false)
-          rawMetaRef.current = meta
-        } catch (e) {
-          console.error(e)
-          setYamlError(true)
-        }
-      })
-    }
-    md.use(markdownItTaskLists, { lineNumber: true })
-    if (plantumlServer) {
-      md.use(plantuml, {
-        openMarker: '```plantuml',
-        closeMarker: '```',
-        server: plantumlServer
-      })
-    } else {
-      md.use(plantumlError)
-    }
-
-    if (onMetaDataChange) {
-      md.use(frontmatter, (rawMeta: string) => {
-        try {
-          const meta: RawYAMLMetadata = yaml.safeLoad(rawMeta) as RawYAMLMetadata
-          setYamlError(false)
-          rawMetaRef.current = meta
-        } catch (e) {
-          console.error(e)
-          setYamlError(true)
-          rawMetaRef.current = ({} as RawYAMLMetadata)
-        }
-      })
-    }
-    // noinspection CheckTagEmptyBody
-    md.use(anchor, {
-      permalink: true,
-      permalinkBefore: true,
-      permalinkClass: 'heading-anchor text-dark',
-      permalinkSymbol: '<i class="fa fa-link"></i>'
-    })
-    md.use(mathJax({
-      beforeMath: '<app-katex>',
-      afterMath: '</app-katex>',
-      beforeInlineMath: '<app-katex inline>',
-      afterInlineMath: '</app-katex>',
-      beforeDisplayMath: '<app-katex>',
-      afterDisplayMath: '</app-katex>'
-    }))
-    md.use(markdownItRegex, replaceLegacyYoutubeShortCode)
-    md.use(markdownItRegex, replaceLegacyVimeoShortCode)
-    md.use(markdownItRegex, replaceLegacyGistShortCode)
-    md.use(markdownItRegex, replaceLegacySlideshareShortCode)
-    md.use(markdownItRegex, replaceLegacySpeakerdeckShortCode)
-    md.use(markdownItRegex, replacePdfShortCode)
-    md.use(markdownItRegex, replaceAsciinemaLink)
-    md.use(markdownItRegex, replaceYouTubeLink)
-    md.use(markdownItRegex, replaceVimeoLink)
-    md.use(markdownItRegex, replaceGistLink)
-    md.use(highlightedCode)
-    md.use(markdownItRegex, replaceQuoteExtraAuthor)
-    md.use(markdownItRegex, replaceQuoteExtraColor)
-    md.use(markdownItRegex, replaceQuoteExtraTime)
-    md.use(toc, {
-      placeholder: '(\\[TOC\\]|\\[toc\\])',
-      listType: 'ul',
-      level: [1, 2, 3],
-      callback: (code: string, ast: TocAst): void => {
-        tocAst.current = ast
-      },
-      slugify: slugify
-    })
-    validAlertLevels.forEach(level => {
-      md.use(markdownItContainer, level, { render: createRenderContainer(level) })
-    })
-    md.use(lineNumberMarker(), {
-      postLineMarkers: (lineMarkers) => {
-        currentLineMarkers.current = lineMarkers
-      }
-    })
-  }, [onMetaDataChange, plantumlServer])
+  const markdownIt = useMemo(() => {
+    return fullMarkdownRender(
+      !!onMetaDataChange,
+      error => setYamlError(error),
+      rawMeta => { rawMetaRef.current = rawMeta },
+      toc => { tocAst.current = toc },
+      lineMarkers => { currentLineMarkers.current = lineMarkers }
+    )
+  }, [onMetaDataChange])
 
   const clearMetadata = useCallback(() => {
     rawMetaRef.current = undefined
@@ -238,7 +133,7 @@ export const FullMarkdownRenderer: React.FC<FullMarkdownRendererProps & Addition
         </Alert>
       </ShowIf>
       <BasicMarkdownRenderer className={className} wide={wide} content={content} componentReplacers={allReplacers}
-        onConfigureMarkdownIt={configureMarkdownIt} documentReference={documentElement}
+        markdownIt={markdownIt} documentReference={documentElement}
         onBeforeRendering={clearMetadata}/>
     </div>
   )
